refactor: ScheduleSearch요청/응답 구조체 폴더 이동
